+++ app/Providers/EventServiceProvider.php
@@ -6,8 +6,6 @@ use Illuminate\Auth\Events\Registered;
 use Illuminate\Auth\Listeners\SendEmailVerificationNotification;
 use Illuminate\Foundation\Support\Providers\EventServiceProvider as ServiceProvider;
 use Illuminate\Support\Facades\Event;
-use Laravel\Passport\Events\AccessTokenCreated;
-use App\Listeners\RevokeOldTokens;
 
 class EventServiceProvider extends ServiceProvider
 {
@@ -18,7 +16,7 @@ class EventServiceProvider extends ServiceProvider
      */
     protected $listen = [
         Registered::class => [
-            SendEmailVerificationNotification::class
+            SendEmailVerificationNotification::class,
         ],
         AccessTokenCreated::class => [
             RevokeOldTokens::class
@@ -32,8 +30,6 @@ class EventServiceProvider extends ServiceProvider
      */
     public function boot()
     {
-        parent::boot();
-
-        //
+        // parent::boot();
     }
 }