+++ database/seeders/RoleUserTableSeeder.php
@@ -1,8 +1,11 @@
 <?php
+
+namespace Database\Seeders;
+
 use Illuminate\Database\Seeder;
 use App\Models\Role;
 use Illuminate\Support\Facades\DB;
-use App\User;
+use App\Models\User;
 
 class RoleUserTableSeeder extends Seeder
 {